backend/requirements.txt
Flask
requests
python-dotenv
markdown
google.generativeai